# Optimize JSON parsing