--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -33,7 +33,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        go: ["1.20", "1.21"]
+        go: ["1.21", "stable"]
     steps:
       - uses: actions/checkout@v3
       - name: Setup go
@@ -80,7 +80,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        go: ["1.20", "1.21"]
+        go: ["1.21", "stable"]
     steps:
       - uses: actions/checkout@v3
         with: